+++ sql/catalyst/src/main/scala/org/apache/spark/sql/internal/SQLConf.scala
@@ -3008,14 +3008,14 @@ object SQLConf {
       .booleanConf
       .createWithDefault(false)
 
-  val PYTHON_TABLE_UDF_ANALYZER_MEMORY =
-    buildConf("spark.sql.analyzer.pythonUDTF.analyzeInPython.memory")
-      .doc("The amount of memory to be allocated to PySpark for Python UDTF analyzer, in MiB " +
-        "unless otherwise specified. If set, PySpark memory for Python UDTF analyzer will be " +
-        "limited to this amount. If not set, Spark will not limit Python's " +
-        "memory use and it is up to the application to avoid exceeding the overhead memory space " +
-        "shared with other non-JVM processes.\nNote: Windows does not support resource limiting " +
-        "and actual resource is not limited on MacOS.")
+  val PYTHON_PLANNER_EXEC_MEMORY =
+    buildConf("spark.sql.planner.pythonExecution.memory")
+      .doc("Specifies the memory allocation for executing Python code in Spark driver, in MiB. " +
+        "When set, it caps the memory for Python execution to the specified amount. " +
+        "If not set, Spark will not limit Python's memory usage and it is up to the application " +
+        "to avoid exceeding the overhead memory space shared with other non-JVM processes.\n" +
+        "Note: Windows does not support resource limiting and actual resource is not limited " +
+        "on MacOS.")
       .version("4.0.0")
       .bytesConf(ByteUnit.MiB)
       .createOptional
@@ -5157,7 +5157,7 @@ class SQLConf extends Serializable with Logging with SqlApiConf {
   def pysparkWorkerPythonExecutable: Option[String] =
     getConf(SQLConf.PYSPARK_WORKER_PYTHON_EXECUTABLE)
 
-  def pythonUDTFAnalyzerMemory: Option[Long] = getConf(PYTHON_TABLE_UDF_ANALYZER_MEMORY)
+  def pythonPlannerExecMemory: Option[Long] = getConf(PYTHON_PLANNER_EXEC_MEMORY)
 
   def replaceExceptWithFilter: Boolean = getConf(REPLACE_EXCEPT_WITH_FILTER)
